+++ react/components/CodeConfirmation.js
@@ -62,6 +62,7 @@ class CodeConfirmation extends Component {
       onStateChange,
       previous,
       code,
+      accessCodePlaceholder,
     } = this.props
     const {
       isLoading,
@@ -76,7 +77,9 @@ class CodeConfirmation extends Component {
         onSubmit={e => this.handleOnSubmit(e)}
         content={(
           <React.Fragment>
-            <Input value={code} onChange={this.handleInputChange} />
+            <div className="vtex-login__input-container vtex-login__input-container--access-code">
+              <Input value={code} onChange={this.handleInputChange} placeholder={accessCodePlaceholder} />
+            </div>
             <FormError show={isInvalidCode}>
               {translate('login.invalidCode', intl)}
             </FormError>
@@ -126,6 +129,8 @@ CodeConfirmation.propTypes = {
   accessKeySignIn: PropTypes.func.isRequired,
   /** Intl object*/
   intl: intlShape,
+  /** Placeholder to access code input */
+  accessCodePlaceholder: PropTypes.string,
   /** Function called after login success */
   loginCallback: PropTypes.func,
 }