+++ react/components/EmailVerification.js
@@ -25,6 +25,8 @@ class EmailVerification extends Component {
     isCreatePassword: PropTypes.bool.isRequired,
     /** Title to be displayed */
     title: PropTypes.string,
+    /** Placeholder to email input */
+    emailPlaceholder: PropTypes.string,
     /** Email set on state */
     email: PropTypes.string.isRequired,
     /** Function to change de active tab */
@@ -82,6 +84,7 @@ class EmailVerification extends Component {
       email,
       isCreatePassword,
       showBackButton,
+      emailPlaceholder,
     } = this.props
     const { isLoading, isInvalidEmail, isUserBlocked } = this.state
 
@@ -92,11 +95,13 @@ class EmailVerification extends Component {
         onSubmit={e => this.handleOnSubmit(e)}
         content={(
           <React.Fragment>
-            <Input
-              value={email}
-              onChange={this.handleInputChange}
-              placeholder={'Ex: example@mail.com'}
-            />
+            <div className="vtex-login__input-container vtex-login__input-container--email">
+              <Input
+                value={email}
+                onChange={this.handleInputChange}
+                placeholder={emailPlaceholder}
+              />
+            </div>
             <FormError show={isInvalidEmail}>
               {translate('login.invalidEmail', intl)}
             </FormError>