+++ CHANGELOG.md
@@ -6,11 +6,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `LoginContent` position to be always under the `ProfileIcon`.
+
+## [0.9.12] - 2018-07-24
+### Fixed
+- Fix onClick handle of recovery password.
+
+## [0.9.11] - 2018-07-24
+### Added
+- Now, placeholders are customize by schema props.
+
+### Changed
+- Allow breakline on Login Titles.
+
+## [0.9.10] - 2018-07-24
 ### Added
 - Support for attribute `closeonclick` to close on click events inside modal.
 
-### Fixed
-- `LoginContent` position to be always under the `ProfileIcon`.
+### Changed
+- Update CCS classes to improve customization.
 
 ## [0.9.9] - 2018-07-24
 ### Fixed